requirements.txt
streamlit
pandas
numpy
openpyxl
xlsxwriter
openai
python-dotenv
plotly
python-dateutil
